[Fix] #44 - LocationManager와 LocationHelper 병합
